refactor(generic,apis_entities): move duplicate logic to GenericModel
The logic for duplicating and the corresponding routes were still part
of the `apis_entities` app. This commit moves them to the generic app,
thus they can also be used for relations and any other entities that
inherit from GenericModel.
