+++ apis_core/generic/abc.py
@@ -51,6 +51,10 @@ class GenericModel:
         ct = ContentType.objects.get_for_model(self)
         return reverse("apis_core:generic:update", args=[ct, self.id])
 
+    def get_duplicate_url(self):
+        ct = ContentType.objects.get_for_model(self)
+        return reverse("apis_core:generic:duplicate", args=[ct, self.id])
+
     def get_enrich_url(self):
         ct = ContentType.objects.get_for_model(self)
         return reverse("apis_core:generic:enrich", args=[ct, self.id])
