+++ apis_core/generic/urls.py
@@ -49,6 +49,7 @@ urlpatterns = [
                 path("create", views.Create.as_view(), name="create"),
                 path("delete/<int:pk>", views.Delete.as_view(), name="delete"),
                 path("update/<int:pk>", views.Update.as_view(), name="update"),
+                path("duplicate/<int:pk>", views.Duplicate.as_view(), name="duplicate"),
                 path(
                     "selectmergeorenrich/<int:pk>",
                     views.SelectMergeOrEnrich.as_view(),
